114 create ranged slider demos
